requirements.txt
astroid==2.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-axes==5.12.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-import-export==2.4.0
django-ipware==3.0.2
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pkg-resources==0.0.0
pylint==2.6.0
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
